Fix MATRIX strobeCount off-by-one, widen CoverGateDef delays, add null guard
- InputControl.cpp: Fix strobeCount calculation when no valid strobes exist
  (was always 1, now correctly 0)
- CoverGateDef.h: Widen delay_ms/close_delay_ms from uint16_t to uint32_t
- DCSBIOSBridge.cpp: Add defensive null check for ae->entries[i]
- Regenerated IFEI label set data

Verified: both MAIN and IFEI compile clean on ESP32-S3.

+++ src/Core/CoverGateDef.h
@@ -15,6 +15,6 @@ struct CoverGateDef {
     const char* release_label;
     const char* cover_label;
     CoverGateKind kind;
-    uint16_t delay_ms;
-    uint16_t close_delay_ms;
+    uint32_t delay_ms;
+    uint32_t close_delay_ms;
 };
